move Dns messages to message
